requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pydantic<2
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-jose[cryptography]
jinja2
python-multipart
pytest
pytesseract
easyocr
Pillow
pdf2image
pypdf
python-dateutil
yfinance
requests
selenium
webdriver-manager
cryptography
plaid-python
